requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
scikit-surprise==1.1.4
scipy==1.15.0
joblib==1.4.2
httpx==0.28.1
python-dotenv==1.0.1
